Don't generate tags on cursorhold
This was getting distracting, as the amount of time it takes to generate
tags is noticeably laggy when done within Vim. The author of easytags
claims he is looking into making it work in a background process, but
until that happens the ability to generate tags on a per-file basis can
be avoided (at the cost of not having dynamically updated tags while in
the middle of editing--an acceptable trade-off).