Add Qargs command
The Qargs command takes all files in the Quickfix window and loads them
into the args list. This allows you to search for a string using the Ack
plugin to get it into the quickfix window, and then you can run Qargs
and iterate over the files in the list using :argdo. This makes
search/replace over multiple files easy.