Change update script to reset --hard
When running `git pull origin master` on the vim-powerline plugin, a
full rebase was being run (probably due to how my git config is set up
to auto-rebase).

Rather than leave this decision up to the user's configuration, change
the behaviour to explicity fetch and then reset the repo's HEAD to
`origin/master`, so that no funny business occurs. This has the added
benefit of producing less output, which is nice.