Don't fetch origin if HEAD matches plugin revision
Previously, if you ran the `update` script and a plugin was defined with
an explicit hash, it would still attempt to fetch changes from the
plugin's upstream origin.

Since having an explicit hash defined for a plugin means that this fetch
never needs to happen once the plugin's `HEAD` has that hash, this
commit optimizes the update script to skip the fetch step in the event
the revision hash is defined and it matches the current plugin's `HEAD`.