Add extra <CR> to some shortcut commands
This prevents the annoying "Press ENTER or type command to continue"
prompt from appearing when these commands are executed in an empty Vim
window.

We can't just disable the prompt because it is useful in other
situations, for example warning before closing a file that has been
modified.

Thus we add an extra <CR> to the problematic shortcuts so that the
message is auto-dismissed.