Add shortcuts for opening associated files
The plugin a.vim allows for the easy opening of files. So easy, in fact,
that making it easier would be awesome.

This adds a couple of shortcuts so that opening an associated file in
the same window, vertical split, horizontal split, or new tab is a
simple operation, and matches up with similar keyboard shortcuts for
other vertical/horizontal related actions.