Fix background colour-filling when in tmux
When using vim within tmux, the background at the end of lines of code
would be filled with the incorrect background colour.

Investigation revealed that this was caused by the `t_Co` setting
changing when the `$TERM` environment variable was not set to
`screen-256color`. Unfortunately, setting `$TERM` to this also disables
keyboard shortcuts involving the Ctrl/Shift/Alt keys, which was
undesirable.

The solution is to clear the `t_ut` variable as described in:

  http://snk.tuxfamily.org/log/vim-256color-bce.html