Move VI incompatibility setting into init.vim
This needs to be the first thing run in order to prevent unexpected
behaviour from occurring in other configuration settings. For example,
setting the `nocompatible` option after setting the `shortmess` option
causes the initial `shortmess` setting to be ignored, which is weird.

Moving this to be the first statement executed saves a lot of potential
headache down the road.