Improve update script to work with tag/branch names
In order to speed up the `update` script, being able to specify a
specific commit or tag name allows us to avoid fetching from the origin
when not necessary.

However, we still always want to fetch when the name is a branch (since
branches can be updated remotely, but tags in practice shouldn't).